+++ pkg/util/tests/logging.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,7 +49,7 @@ func (l *logScanner) GetData(t *testing.T, timeout time.Duration, obj interface{
 	if s, ok := l.Get(timeout); !ok {
 		return false
 	} else {
-		require.NoError(t, json.Unmarshal([]byte(s), obj))
+		require.NoError(t, json.Unmarshal([]byte(s), obj), s)
 		return true
 	}
 }