+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Maintenance) Bump Prometheus API Version
 - (Bugfix) Prevent unexpected rotation in case of SecurityContext change
 - (Bugfix) Ensure PDB is created
+- (Bugfix) Fix Schema Apply Checksum
 
 ## [1.2.40](https://github.com/arangodb/kube-arangodb/tree/1.2.40) (2024-04-10)
 - (Feature) Add Core fields to the Scheduler Container Spec